Refactor: Fix misplaced responsibility.

+++ src/test/java/org/henrysgrocery/store/Item.java
@@ -1,13 +1,15 @@
 package org.henrysgrocery.store;
 
+import java.math.BigDecimal;
+
 public class Item {
     final Unit bottle;
     final String milk;
-    final double price;
+    final BigDecimal price;
 
     public Item(Unit bottle, String milk, double price) {
         this.bottle = bottle;
         this.milk = milk;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 }
